+++ src/app/_component/common/Input/_components/ClearButton.variants.ts
